Revert "* FIX [hashmap] fixed lock error."
This reverts commit 7b7d15431b620a4149c9a264309c5365ad8b34c3.
